feat: renamed unspec to civil (CMC-1541)
