+++ audit.json
@@ -1,7 +1,7 @@
 {
-  "100001_Unexpected Content-Type was returned_http://unspec-service-aat.service.core-compute-aat.internal/_GET":  "ignore",
-  "100000_A Client Error response code was returned by the server_http://unspec-service-aat.service.core-compute-aat.internal/cases/callbacks/version/version/callback-type_POST":  "ignore",
-  "100000_A Client Error response code was returned by the server_http://unspec-service-aat.service.core-compute-aat.internal/cases/callbacks/version/version/callback-type/page-id_POST": "ignore",
-  "100000_A Client Error response code was returned by the server_http://unspec-service-aat.service.core-compute-aat.internal/cases/callbacks/callback-type_POST":  "ignore",
-  "100000_A Client Error response code was returned by the server_http://unspec-service-aat.service.core-compute-aat.internal/cases/callbacks/callback-type/page-id_POST": "ignore"
+  "100001_Unexpected Content-Type was returned_http://civil-service-aat.service.core-compute-aat.internal/_GET":  "ignore",
+  "100000_A Client Error response code was returned by the server_http://civil-service-aat.service.core-compute-aat.internal/cases/callbacks/version/version/callback-type_POST":  "ignore",
+  "100000_A Client Error response code was returned by the server_http://civil-service-aat.service.core-compute-aat.internal/cases/callbacks/version/version/callback-type/page-id_POST": "ignore",
+  "100000_A Client Error response code was returned by the server_http://civil-service-aat.service.core-compute-aat.internal/cases/callbacks/callback-type_POST":  "ignore",
+  "100000_A Client Error response code was returned by the server_http://civil-service-aat.service.core-compute-aat.internal/cases/callbacks/callback-type/page-id_POST": "ignore"
 }
