+++ bin/import-ccd-definition.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 definition_input_dir=$(realpath 'ccd-definition')
-definition_output_file="$(realpath ".")/build/ccd-development-config/ccd-unspec-dev.xlsx"
+definition_output_file="$(realpath ".")/build/ccd-development-config/ccd-civil-dev.xlsx"
 params="$@"
 
 ./bin/utils/import-ccd-definition.sh "${definition_input_dir}" "${definition_output_file}" "${params}"
