+++ build.gradle
@@ -124,7 +124,7 @@ task checkDependenciesIntegrity(type: Exec, description: 'Checks integrity of Ya
 
 task awaitApplicationReadiness(type: Exec, description: 'Awaits until application is ready.') {
   commandLine './bin/wait-for.sh', System.env.URL
-  commandLine './bin/wait-for.sh', System.env.UNSPEC_SERVICE_URL
+  commandLine './bin/wait-for.sh', System.env.CIVIL_SERVICE_URL
 }
 
 task runSmokeTests(type: Exec, description: 'Runs smoke tests.') {
@@ -212,7 +212,7 @@ dependencies {
   testImplementation group: 'com.h2database', name: 'h2', version: '1.4.200'
 }
 
-mainClassName = 'uk.gov.hmcts.reform.unspec.Application'
+mainClassName = 'uk.gov.hmcts.reform.civil.Application'
 
 bootJar {
   getArchiveFileName().set(provider {
