+++ charts/civil-camunda/Chart.yaml
@@ -14,7 +14,7 @@ dependencies:
     version: 5.0.2
     repository: '@hmctspublic'
     tags:
-      - unspec-ccd-stack
+      - civil-ccd-stack
   - name: xui-webapp
     version: ~0.2.1
     repository: '@hmctspublic'
