+++ charts/civil-camunda/values.aat.template.yaml
@@ -1,5 +1,5 @@
 tags:
-  unspec-ccd-stack: true
+  civil-ccd-stack: true
 
 postgresql:
   enabled: true
@@ -26,16 +26,16 @@ postgresql:
 
 java:
   applicationPort: 4000
-  aadIdentityName: unspec
+  aadIdentityName: civil
   image: 'hmctspublic.azurecr.io/civil/service:latest'
   imagePullPolicy: Always
   ingressHost: ${SERVICE_FQDN}
   keyVaults:
-    unspec:
-      resourceGroup: unspec
+    civil:
+      resourceGroup: civil
       secrets:
-        - unspec-idam-client-secret
-        - microservicekey-unspec-service
+        - civil-idam-client-secret
+        - microservicekey-civil-service
         - system-update-user-username
         - system-update-user-password
         - cross-access-user-username
@@ -50,7 +50,7 @@ java:
         - robotics-notification-recipient
         - launch-darkly-sdk-key
   environment:
-    IDAM_CLIENT_ID: unspec-service
+    IDAM_CLIENT_ID: civil-service
     IDAM_CLIENT_REDIRECT_URI: https://civil-camunda-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/oauth2/callback
     IDAM_API_URL: https://idam-api.{{ .Values.global.environment }}.platform.hmcts.net
     IDAM_WEB_URL: https://idam-web-public.{{ .Values.global.environment }}.platform.hmcts.net
@@ -149,10 +149,10 @@ ccd:
       environment:
         DATA_STORE_DB_HOST: ${SERVICE_NAME}-postgresql
         DATA_STORE_IDAM_KEY: ${CCD_DATA_STORE_S2S_SECRET}
-        DATA_STORE_S2S_AUTHORISED_SERVICES: ccd_data,ccd_gw,ccd_ps,bulk_scan_orchestrator,ccpay_bubble,ctsc_work_allocation,em_ccd_orchestrator,xui_webapp,unspec_service
+        DATA_STORE_S2S_AUTHORISED_SERVICES: ccd_data,ccd_gw,ccd_ps,bulk_scan_orchestrator,ccpay_bubble,ctsc_work_allocation,em_ccd_orchestrator,xui_webapp,civil_service
         IDAM_USER_URL: https://idam-web-public.aat.platform.hmcts.net
         OIDC_ISSUER: https://forgerock-am.service.core-compute-idam-aat2.internal:8443/openam/oauth2/realms/root/realms/hmcts
-        CCD_S2S_AUTHORISED_SERVICES_CASE_USER_ROLES: unspec_service
+        CCD_S2S_AUTHORISED_SERVICES_CASE_USER_ROLES: civil_service
       keyVaults: []
       ingressHost: civil-camunda-data-store-staging-aat.service.core-compute-aat.internal
   ccd-definition-store-api:
