+++ charts/civil-camunda/values.yaml
@@ -1,5 +1,5 @@
 tags:
-  unspec-ccd-stack: false
+  civil-ccd-stack: false
 
 postgresql:
   enabled: false
@@ -16,13 +16,13 @@ camunda-bpm:
 java:
   applicationPort: 4000
   image: 'hmctspublic.azurecr.io/civil/service:latest'
-  aadIdentityName: unspec
+  aadIdentityName: civil
   keyVaults:
-    unspec:
-      resourceGroup: unspec
+    civil:
+      resourceGroup: civil
       secrets:
-        - unspec-idam-client-secret
-        - microservicekey-unspec-service
+        - civil-idam-client-secret
+        - microservicekey-civil-service
         - system-update-user-username
         - system-update-user-password
         - AppInsightsInstrumentationKey
