+++ src/test/java/uk/gov/hmcts/reform/civil/bpmn/AcknowledgeClaimTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.unspec.bpmn;
+package uk.gov.hmcts.reform.civil.bpmn;
 
 import org.camunda.bpm.engine.externaltask.ExternalTask;
 import org.camunda.bpm.engine.variable.VariableMap;
@@ -8,20 +8,22 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-class TakeCaseOfflineTest extends BpmnBaseTest {
+class AcknowledgeClaimTest extends BpmnBaseTest {
 
-    public static final String MESSAGE_NAME = "TAKE_CASE_OFFLINE";
-    public static final String PROCESS_ID = "TAKE_CASE_OFFLINE";
+    private static final String MESSAGE_NAME = "ACKNOWLEDGE_CLAIM";
+    private static final String PROCESS_ID = "ACKNOWLEDGE_CLAIM_PROCESS_ID";
 
-    private static final String NOTIFY_RPA_ON_CASE_HANDED_OFFLINE = "NOTIFY_RPA_ON_CASE_HANDED_OFFLINE";
-    private static final String NOTIFY_RPA_ON_CASE_HANDED_OFFLINE_ACTIVITY_ID = "NotifyRoboticsOnCaseHandedOffline";
+    private static final String NOTIFY_APPLICANT_SOLICITOR_1 = "NOTIFY_APPLICANT_SOLICITOR1_FOR_CLAIM_ACKNOWLEDGEMENT";
+    private static final String GENERATE_ACKNOWLEDGEMENT_OF_CLAIM = "GENERATE_ACKNOWLEDGEMENT_OF_CLAIM";
+    private static final String NOTIFICATION_ACTIVITY_ID = "AcknowledgeClaimNotifyApplicantSolicitor1";
+    private static final String GENERATE_CERTIFICATE_ACTIVITY_ID = "AcknowledgeClaimGenerateAcknowledgementOfClaim";
 
-    public TakeCaseOfflineTest() {
-        super("take_case_offline.bpmn", "TAKE_CASE_OFFLINE");
+    public AcknowledgeClaimTest() {
+        super("acknowledge_claim.bpmn", "ACKNOWLEDGE_CLAIM_PROCESS_ID");
     }
 
     @Test
-    void shouldSuccessfullyCompleteTakeCaseOffline() {
+    void shouldSuccessfullyCompleteAcknowledgeClaim() {
         //assert process has started
         assertFalse(processInstance.isEnded());
 
@@ -32,14 +34,15 @@ class TakeCaseOfflineTest extends BpmnBaseTest {
         ExternalTask startBusiness = assertNextExternalTask(START_BUSINESS_TOPIC);
         assertCompleteExternalTask(startBusiness, START_BUSINESS_TOPIC, START_BUSINESS_EVENT, START_BUSINESS_ACTIVITY);
 
-        //complete the RPA notification
-        ExternalTask respondentNotification = assertNextExternalTask(PROCESS_CASE_EVENT);
-        assertCompleteExternalTask(
-            respondentNotification,
-            PROCESS_CASE_EVENT,
-            NOTIFY_RPA_ON_CASE_HANDED_OFFLINE,
-            NOTIFY_RPA_ON_CASE_HANDED_OFFLINE_ACTIVITY_ID
-        );
+        //complete the document generation
+        ExternalTask documentGeneration = assertNextExternalTask(PROCESS_CASE_EVENT);
+        assertCompleteExternalTask(documentGeneration, PROCESS_CASE_EVENT, GENERATE_ACKNOWLEDGEMENT_OF_CLAIM,
+                                   GENERATE_CERTIFICATE_ACTIVITY_ID);
+
+        //complete the notification to applicant
+        ExternalTask notification = assertNextExternalTask(PROCESS_CASE_EVENT);
+        assertCompleteExternalTask(notification, PROCESS_CASE_EVENT, NOTIFY_APPLICANT_SOLICITOR_1,
+                                   NOTIFICATION_ACTIVITY_ID);
 
         //end business process
         ExternalTask endBusinessProcess = assertNextExternalTask(END_BUSINESS_PROCESS);
