+++ src/test/java/uk/gov/hmcts/reform/civil/bpmn/TakeCaseOfflineSchedulerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.unspec.bpmn;
+package uk.gov.hmcts.reform.civil.bpmn;
 
 import org.camunda.bpm.engine.externaltask.ExternalTask;
 import org.camunda.bpm.engine.externaltask.LockedExternalTask;
@@ -10,16 +10,16 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-class NotifyClaimDeadlineSchedulerTest extends BpmnBaseTest {
+class TakeCaseOfflineSchedulerTest extends BpmnBaseTest {
 
-    public static final String TOPIC_NAME = "CASE_DISMISSED";
+    public static final String TOPIC_NAME = "TAKE_CASE_OFFLINE";
 
-    public NotifyClaimDeadlineSchedulerTest() {
-        super("notify_claim_deadline_scheduler.bpmn", "NOTIFY_CLAIM_DEADLINE_SCHEDULER");
+    public TakeCaseOfflineSchedulerTest() {
+        super("take_case_offline_scheduler.bpmn", "TAKE_CASE_OFFLINE_SCHEDULER");
     }
 
     @Test
-    void notifyClaimDeadlineSchedulerShouldFireCaseDismissedExternalTask_whenStarted() {
+    void schedulerShouldRaiseTakeCaseOfflineExternalTask_whenStarted() {
         //assert process has started
         assertFalse(processInstance.isEnded());
 
@@ -33,7 +33,7 @@ class NotifyClaimDeadlineSchedulerTest extends BpmnBaseTest {
         assertThat(jobDefinitions).hasSize(1);
         assertThat(jobDefinitions.get(0).getJobType()).isEqualTo("timer-start-event");
 
-        assertThat(jobDefinitions.get(0).getJobConfiguration()).isEqualTo("CYCLE: 5 0 * * * ?");
+        assertThat(jobDefinitions.get(0).getJobConfiguration()).isEqualTo("CYCLE: 5 16 * * * ?");
 
         //get external tasks
         List<ExternalTask> externalTasks = getExternalTasks();
