+++ src/test/java/uk/gov/hmcts/reform/civil/bpmn/TakeCaseOfflineTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.unspec.bpmn;
+package uk.gov.hmcts.reform.civil.bpmn;
 
 import org.camunda.bpm.engine.externaltask.ExternalTask;
 import org.camunda.bpm.engine.variable.VariableMap;
@@ -8,22 +8,20 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-class DismissClaimTest extends BpmnBaseTest {
+class TakeCaseOfflineTest extends BpmnBaseTest {
 
-    public static final String MESSAGE_NAME = "DISMISS_CLAIM";
-    public static final String PROCESS_ID = "DISMISS_CLAIM";
+    public static final String MESSAGE_NAME = "TAKE_CASE_OFFLINE";
+    public static final String PROCESS_ID = "TAKE_CASE_OFFLINE";
 
-    public static final String NOTIFY_RESPONDENT_SOLICITOR_1 = "NOTIFY_RESPONDENT_SOLICITOR1_CLAIM_DISMISSED";
-    public static final String RESPONDENT_SOLICITOR_1_ACTIVITY_ID = "ClaimDismissedNotifyRespondentSolicitor1";
-    public static final String NOTIFY_APPLICANT_SOLICITOR_1 = "NOTIFY_APPLICANT_SOLICITOR1_CLAIM_DISMISSED";
-    public static final String APPLICANT_SOLICITOR_1_ACTIVITY_ID = "ClaimDismissedNotifyApplicantSolicitor1";
+    private static final String NOTIFY_RPA_ON_CASE_HANDED_OFFLINE = "NOTIFY_RPA_ON_CASE_HANDED_OFFLINE";
+    private static final String NOTIFY_RPA_ON_CASE_HANDED_OFFLINE_ACTIVITY_ID = "NotifyRoboticsOnCaseHandedOffline";
 
-    public DismissClaimTest() {
-        super("claim_dismissed.bpmn", "DISMISS_CLAIM");
+    public TakeCaseOfflineTest() {
+        super("take_case_offline.bpmn", "TAKE_CASE_OFFLINE");
     }
 
     @Test
-    void shouldSuccessfullyCompleteDismissClaim() {
+    void shouldSuccessfullyCompleteTakeCaseOffline() {
         //assert process has started
         assertFalse(processInstance.isEnded());
 
@@ -34,22 +32,13 @@ class DismissClaimTest extends BpmnBaseTest {
         ExternalTask startBusiness = assertNextExternalTask(START_BUSINESS_TOPIC);
         assertCompleteExternalTask(startBusiness, START_BUSINESS_TOPIC, START_BUSINESS_EVENT, START_BUSINESS_ACTIVITY);
 
-        //complete the notification to respondent solicitor 1
+        //complete the RPA notification
         ExternalTask respondentNotification = assertNextExternalTask(PROCESS_CASE_EVENT);
         assertCompleteExternalTask(
             respondentNotification,
             PROCESS_CASE_EVENT,
-            NOTIFY_RESPONDENT_SOLICITOR_1,
-            RESPONDENT_SOLICITOR_1_ACTIVITY_ID
-        );
-
-        //complete the notification to applicant solicitor 1
-        ExternalTask applicantNotification = assertNextExternalTask(PROCESS_CASE_EVENT);
-        assertCompleteExternalTask(
-            applicantNotification,
-            PROCESS_CASE_EVENT,
-            NOTIFY_APPLICANT_SOLICITOR_1,
-            APPLICANT_SOLICITOR_1_ACTIVITY_ID
+            NOTIFY_RPA_ON_CASE_HANDED_OFFLINE,
+            NOTIFY_RPA_ON_CASE_HANDED_OFFLINE_ACTIVITY_ID
         );
 
         //end business process
